Move version command into the cmd package like the rest
Longhorn #1983
